Create constants file and change to station based calculation
Constants are kept in a global classdef so that they may be accessed
uniformly everywhere in the file. All station location information goes
into the cell array called all_pts_info, from which other constants are
declared from.

Create an arbitrarily oriented prism calculation based on the
orientation of the plane. This is given as three basis vectors which
describe the relatively oriented Euclidean basis of the prism.