Use a map for the point database instead of all_pts_info
Correct names in point_database.csv to match up with those in
gz_measurements. This simplifies the constants class. A table is used
because it can be indexed by row name (station name, effectively) just
like a map, but has the added benefit of having multiple column variable
instead of just a single value.